ROFO-228 refactor: 공통화면 common으로 분리
